12/08 버그 수정
recvBuf 버퍼 사이즈가 작아서 모든 데이터를 다 받아서 처리하지 못하는 문제
-> 버퍼 사이즈를 늘려서 해결
->int short 자료형이 혼재해 생기는 오류 수정

+++ PerfectShape_Client_/PerfectShape_Client_/main.cpp
@@ -265,8 +265,8 @@ clock_t start;
 
 int myID;
 struct NetPlayer {
-	short id;
-	short hp = 3 ;
+	int id;
+	int hp = 3 ;
 	GLUquadricObj* ball;
 	glm::vec3 t;
 	bool active = true;
